stocks_app/requirements.txt
python-dotenv
django
djangorestframework
django-filter
psycopg2-binary
